+++ src/api/players/controllers/deletePlayer.controller.ts
@@ -1,16 +1,20 @@
 import type { NextFunction, Response, Request } from 'express';
 
 import { deletePlayerService } from '../services/deletePlayer.service';
+import createHttpError from 'http-errors';
 
 export async function deletePlayerController(
   req: Request,
   res: Response,
-  _next: NextFunction,
+  next: NextFunction,
 ) {
+  if (!req.body?.user?.isAdmin) {
+    return next(createHttpError(403, 'Only admins can delete players'));
+  }
+
   await deletePlayerService(req.params.id);
 
   res.json({
     message: 'Player deleted successfully',
   });
 }
-
